+++ infrastructure/file/src/main/java/cmc/mellyserver/FileDto.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.file;
+package cmc.mellyserver;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
