+++ infrastructure/notification/src/main/java/cmc/mellyserver/email/javamail/JavaMailSendService.java
@@ -1,6 +1,6 @@
-package cmc.mellyserver.notification.email.javamail;
+package cmc.mellyserver.email.javamail;
 
-import cmc.mellyserver.notification.email.EmailSendService;
+import cmc.mellyserver.email.EmailSendService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
