+++ infrastructure/notification/src/main/java/cmc/mellyserver/message/MessageService.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.notification.fcm;
+package cmc.mellyserver.message;
 
 
 public interface MessageService {
@@ -6,6 +6,4 @@ public interface MessageService {
     void sendCommentCreatedMessage(final Long userId, final Long memoryId, final String nickname);
 
     void sendCommentLikeCreatedMessage(final Long userId, final Long memoryId, final String nickname);
-
-    void sendGroupUserCreatedMemoryMessage(final Long userId, final Long memoryId, final String nickname);
 }
