+++ infrastructure/notification/src/main/java/cmc/mellyserver/message/constants/NotificationConstants.java
@@ -1,0 +1,14 @@
+package cmc.mellyserver.message.constants;
+
+public abstract class NotificationConstants {
+
+    public static final String COMMENT_CREATED_TITLE = "댓글 생성 알림";
+
+    public static final String COMMENT_CREATED_CONTENT = "님이 댓글을 달았습니다";
+
+    public static final String COMMENT_LIKE_NOTI_TITLE = "댓글 좋아요 알림";
+
+    public static final String COMMENT_LIKE_NOTI_CONTENT = "님이 회원님의 댓글을 좋아합니다";
+
+    public static final String GROUP_USER_CREATED_MEMORY_CONTENT = "님이 메모리를 올렸습니다. 한번 확인해보세요!";
+}
