--- infrastructure/notification/src/main/java/cmc/mellyserver/notification/constants/NotificationConstants.java
@@ -1,8 +1,0 @@
-package cmc.mellyserver.notification.constants;
-
-public abstract class NotificationConstants {
-
-    public static final String COMMENT_LIKE_NOTI_CONTENT = "님이 회원님의 댓글을 좋아합니다";
-
-    public static final String GROUP_USER_CREATED_MEMORY_CONTENT = "님이 메모리를 올렸습니다. 한번 확인해보세요!";
-}
