--- storage/db-redis/src/main/java/cmc/mellyserver/dbredis/DbRedisApplication.java
@@ -1,13 +1,0 @@
-package cmc.mellyserver.dbredis;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DbRedisApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(DbRedisApplication.class, args);
-    }
-
-}
