+++ storage/db-redis/src/main/java/cmc/mellyserver/dbredis/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package cmc.mellyserver.dbredis.redis;
+package cmc.mellyserver.dbredis.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
